Backends/Queue_system/pyproject.toml
[project]
name = "queue-system"
version = "0.1.0"
description = ""
authors = [
    {name = "kikiru328",email = "kikiru328@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "redis[asyncio] (>=6.2.0,<7.0.0)",
    "fastapi-utils (>=0.8.0,<0.9.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "typing-inspect (>=0.9.0,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
